backend/requirements.txt
fastapi
uvicorn
paho-mqtt
python-dotenv
google-generativeai
